#133 - Collapsing all *CronConfig classes into an inner classes of *
