+++ src/main/java/dev/iakunin/codexiabot/bot/config/ForksUpCronConfig.java
@@ -1,7 +1,12 @@
 package dev.iakunin.codexiabot.bot.config;
 
 import dev.iakunin.codexiabot.bot.Up;
+import dev.iakunin.codexiabot.bot.repository.ForksUpResultRepository;
+import dev.iakunin.codexiabot.bot.up.Forks;
+import dev.iakunin.codexiabot.codexia.CodexiaModule;
 import dev.iakunin.codexiabot.common.runnable.Logging;
+import dev.iakunin.codexiabot.github.GithubModule;
+import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -35,5 +40,28 @@ public class ForksUpCronConfig implements SchedulingConfigurer {
             this.forksUpRunnable(),
             this.cronExpression
         );
+    }
+
+    @Configuration
+    @AllArgsConstructor
+    public static class ForksUpConfig {
+
+        private final GithubModule github;
+
+        private final CodexiaModule codexia;
+
+        private final ForksUpResultRepository repository;
+
+        private final Forks bot;
+
+        @Bean
+        public Up forksUp() {
+            return new Up(
+                this.github,
+                this.repository,
+                this.bot,
+                this.codexia
+            );
+        }
     }
 }
