+++ src/main/java/dev/iakunin/codexiabot/bot/config/FoundOnHackernewsCronConfig.java
@@ -1,7 +1,12 @@
 package dev.iakunin.codexiabot.bot.config;
 
+import dev.iakunin.codexiabot.bot.Bot;
 import dev.iakunin.codexiabot.bot.Found;
+import dev.iakunin.codexiabot.bot.found.Hackernews;
+import dev.iakunin.codexiabot.codexia.CodexiaModule;
 import dev.iakunin.codexiabot.common.runnable.Logging;
+import dev.iakunin.codexiabot.github.GithubModule;
+import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -35,5 +40,26 @@ public class FoundOnHackernewsCronConfig implements SchedulingConfigurer {
             this.foundOnHackernewsRunnable(),
             this.cronExpression
         );
+    }
+
+    @Configuration
+    @AllArgsConstructor
+    public static class FoundOnHackernewsConfig {
+
+        private final GithubModule githubModule;
+
+        private final CodexiaModule codexiaModule;
+
+        private final Hackernews bot;
+
+        @Bean
+        public Found foundOnHackernews() {
+            return new Found(
+                Bot.Type.FOUND_ON_HACKERNEWS,
+                this.githubModule,
+                this.codexiaModule,
+                this.bot
+            );
+        }
     }
 }
