+++ src/main/java/dev/iakunin/codexiabot/bot/config/FoundOnRedditCronConfig.java
@@ -1,7 +1,12 @@
 package dev.iakunin.codexiabot.bot.config;
 
+import dev.iakunin.codexiabot.bot.Bot;
 import dev.iakunin.codexiabot.bot.Found;
+import dev.iakunin.codexiabot.bot.found.Reddit;
+import dev.iakunin.codexiabot.codexia.CodexiaModule;
 import dev.iakunin.codexiabot.common.runnable.Logging;
+import dev.iakunin.codexiabot.github.GithubModule;
+import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -35,5 +40,26 @@ public class FoundOnRedditCronConfig implements SchedulingConfigurer {
             this.foundOnRedditRunnable(),
             this.cronExpression
         );
+    }
+
+    @Configuration
+    @AllArgsConstructor
+    public static class FoundOnRedditConfig {
+
+        private final GithubModule githubModule;
+
+        private final CodexiaModule codexiaModule;
+
+        private final Reddit bot;
+
+        @Bean
+        public Found foundOnReddit() {
+            return new Found(
+                Bot.Type.FOUND_ON_REDDIT,
+                this.githubModule,
+                this.codexiaModule,
+                this.bot
+            );
+        }
     }
 }
