+++ src/main/java/dev/iakunin/codexiabot/bot/config/NotSmallCronConfig.java
@@ -1,7 +1,11 @@
 package dev.iakunin.codexiabot.bot.config;
 
 import dev.iakunin.codexiabot.bot.Small;
+import dev.iakunin.codexiabot.bot.repository.TooSmallResultRepository;
+import dev.iakunin.codexiabot.codexia.CodexiaModule;
 import dev.iakunin.codexiabot.common.runnable.Logging;
+import dev.iakunin.codexiabot.github.GithubModule;
+import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -35,5 +39,28 @@ public class NotSmallCronConfig implements SchedulingConfigurer {
             this.notSmallRunnable(),
             this.cronExpression
         );
+    }
+
+    @Configuration
+    @AllArgsConstructor
+    public static class NotSmallConfig {
+
+        private final GithubModule github;
+
+        private final CodexiaModule codexia;
+
+        private final TooSmallResultRepository repository;
+
+        private final dev.iakunin.codexiabot.bot.toosmall.NotSmall bot;
+
+        @Bean
+        public Small notSmall() {
+            return new Small(
+                this.github,
+                this.bot,
+                this.codexia,
+                this.repository
+            );
+        }
     }
 }
