+++ src/main/java/dev/iakunin/codexiabot/bot/config/TooManyStarsCronConfig.java
@@ -1,7 +1,11 @@
 package dev.iakunin.codexiabot.bot.config;
 
 import dev.iakunin.codexiabot.bot.TooManyStars;
+import dev.iakunin.codexiabot.bot.repository.TooManyStarsResultRepository;
+import dev.iakunin.codexiabot.codexia.CodexiaModule;
 import dev.iakunin.codexiabot.common.runnable.Logging;
+import dev.iakunin.codexiabot.github.GithubModule;
+import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -34,5 +38,25 @@ public class TooManyStarsCronConfig implements SchedulingConfigurer {
             this.tooManyStarsRunnable(),
             this.cronExpression
         );
+    }
+
+    @Configuration
+    @AllArgsConstructor
+    public static class TooManyStarsConfig {
+
+        private final GithubModule github;
+
+        private final CodexiaModule codexia;
+
+        private final TooManyStarsResultRepository repository;
+
+        @Bean
+        public TooManyStars tooManyStars() {
+            return new TooManyStars(
+                this.github,
+                this.repository,
+                this.codexia
+            );
+        }
     }
 }
